Fix v7.1.4 Transport Sweep mission progress
