+++ .github/scripts/test_issue523_transport_sweep_progress_runtime.js
@@ -1,0 +1,16 @@
+#!/usr/bin/env node
+'use strict';
+const assert=require('node:assert/strict');const fs=require('node:fs');const path=require('node:path');const vm=require('node:vm');
+const root=path.resolve(__dirname,'..','..');const source=fs.readFileSync(path.join(root,'src','MissionChief_Map_Command_Toolkit.user.js'),'utf8');
+function extractFunction(name){const markers=[`    function ${name}(`,`    async function ${name}(`];const starts=markers.map(marker=>source.indexOf(marker)).filter(index=>index>=0);assert.ok(starts.length,`${name} missing`);const start=Math.min(...starts);const signatureEnd=source.indexOf(') {',start);assert.ok(signatureEnd>=0,`${name} signature end missing`);const open=signatureEnd+2;let depth=0,quote='',escaped=false;for(let index=open;index<source.length;index+=1){const char=source[index];if(quote){if(escaped)escaped=false;else if(char==='\\')escaped=true;else if(char===quote)quote='';continue;}if(char==='"'||char==="'"||char==='`'){quote=char;continue;}if(char==='{')depth+=1;if(char==='}'&&--depth===0)return source.slice(start,index+1);}throw new Error(`Could not extract ${name}`);}
+const names=['transportSweepMissionProgress','setTransportSweepMissionProgress','completeTransportSweepMissionProgress','finaliseTransportSweepMissionProgress'];let renders=0;
+const runtime={queue:[{},{},{}],missionIndex:0,missionTotal:0,completedMissionCount:0,currentItem:'',statusMessage:'',statusLevel:'info',cleared:0,processed:0,skipped:0,errors:0};
+const sandbox={console,Math,Number,String,Boolean,Object,transportSweepRuntime:runtime,renderTransportSweepPanel:()=>{renders+=1;}};vm.createContext(sandbox);vm.runInContext(`${names.map(extractFunction).join('\n\n')}\nthis.api={${names.join(',')}};`,sandbox);const api=sandbox.api;
+let progress=api.setTransportSweepMissionProgress(1,3,{item:'Mission one',message:'Processing mission 1 of 3',render:false});assert.deepEqual(JSON.parse(JSON.stringify(progress)),{current:1,total:3,completed:0,text:'1/3'});assert.equal(runtime.currentItem,'Mission one');assert.equal(renders,0);
+runtime.cleared=3;runtime.processed=3;assert.equal(api.transportSweepMissionProgress().text,'1/3','Patient outcomes must not increment mission progress');
+assert.equal(api.completeTransportSweepMissionProgress(1,{forceRender:true}),true);assert.equal(runtime.completedMissionCount,1);assert.equal(api.completeTransportSweepMissionProgress(1,{forceRender:true}),false,'Repeated finalisation must not increment twice');assert.equal(runtime.completedMissionCount,1);
+progress=api.setTransportSweepMissionProgress(2,3,{forceRender:true});assert.equal(progress.text,'2/3');runtime.skipped+=1;assert.equal(api.completeTransportSweepMissionProgress(2,{forceRender:true}),true);assert.equal(runtime.completedMissionCount,2);
+runtime.errors+=1;assert.equal(api.completeTransportSweepMissionProgress(2,{forceRender:true}),false,'Mission-level error reconciliation must remain idempotent');assert.equal(api.transportSweepMissionProgress().text,'2/3');
+progress=api.finaliseTransportSweepMissionProgress(true);assert.equal(progress.text,'2/3','Cancellation must preserve the last accurate mission position');assert.equal(runtime.completedMissionCount,2);
+runtime.missionIndex=1;runtime.missionTotal=3;runtime.completedMissionCount=0;api.completeTransportSweepMissionProgress(1,{render:false});api.setTransportSweepMissionProgress(2,3,{render:false});api.completeTransportSweepMissionProgress(2,{render:false});api.setTransportSweepMissionProgress(3,3,{render:false});api.completeTransportSweepMissionProgress(3,{render:false});progress=api.finaliseTransportSweepMissionProgress(false);assert.deepEqual(JSON.parse(JSON.stringify(progress)),{current:3,total:3,completed:3,text:'3/3'});assert.ok(renders>=5,'Canonical progress changes must drive managed renders');
+console.log('Issue #523 Transport Sweep mission-progress runtime contract passed.');
