--- .github/workflows/diagnose-issue523-transport-sweep.yml
@@ -1,39 +1,0 @@
-name: Diagnose Issue 523 Transport Sweep Progress
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/issue523/extract_transport_sweep.py"
-      - ".github/workflows/diagnose-issue523-transport-sweep.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-523-transport-sweep-diagnostic
-  cancel-in-progress: false
-
-jobs:
-  diagnose:
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-523-transport-sweep-mission-progress'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - name: Check out exact Issue 523 branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-523-transport-sweep-mission-progress
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Extract Transport Sweep runtime and queue boundaries
-        run: python3 .github/issue523/extract_transport_sweep.py
-
-      - name: Upload Issue 523 diagnostic
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue523-transport-sweep-diagnostic-${{ github.sha }}
-          path: issue523-transport-sweep-diagnostic.txt
-          if-no-files-found: error
-          retention-days: 14
