+++ README.md
@@ -13,7 +13,7 @@
 [![Explore interfaces](https://img.shields.io/badge/EXPLORE-7_INTERFACE_SYSTEMS-6D28D9?style=for-the-badge&logo=palette&logoColor=white)](https://conroy1988.github.io/missionchief-toolkit-assets/themes/)
 [![Release status](https://img.shields.io/badge/OPEN-RELEASE_CONTROL_PANEL-111827?style=for-the-badge&logo=githubactions&logoColor=white)](status/README.md)
 
-## **Current verified release: `v7.1.2` · Development candidate: `v7.1.3` — Continuous Incident Command news reel**
+## **Current verified release: `v7.1.3` · Development candidate: `v7.1.4` — Patient Transport Sweep mission-progress synchronization**
 
 [![GitHub release](https://img.shields.io/github/v/release/Conroy1988/missionchief-toolkit-assets?display_name=release&label=RELEASE&color=2563eb)](https://github.com/Conroy1988/missionchief-toolkit-assets/releases/latest)
 [![Greasy Fork](https://img.shields.io/greasyfork/v/586018?label=GREASY%20FORK&color=670000)](https://greasyfork.org/en/scripts/586018-missionchief-map-command-toolkit)
